# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, legal, finance, sales]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
    CreateUserRequest:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, legal, finance, sales]
        password:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, legal, finance, sales]
        isActive:
          type: boolean
        password:
          type: string
...
    RevenueReport:
      type: object
      required: [id, contractId, period, status, createdAt]
      properties:
        id:
          type: string
        contractId:
          type: string
        partnerName:
          type: string
          nullable: true
        period:
          type: string
        expectedDate:
          type: string
          format: date
          nullable: true
        receivedDate:
          type: string
          format: date
          nullable: true
        amount:
          type: number
          nullable: true
        status:
          type: string
          enum: [expected, received, overdue]
        createdAt:
          type: string
          format: date-time
    CreateRevenueReportRequest:
      type: object
      required: [period, status]
      properties:
        period:
          type: string
        expectedDate:
          type: string
          format: date
          nullable: true
        receivedDate:
          type: string
          format: date
          nullable: true
        amount:
          type: number
          nullable: true
        status:
          type: string
          enum: [expected, received, overdue]
    UpdateRevenueReportRequest:
      type: object
      properties:
        period:
          type: string
        expectedDate:
          type: string
          format: date
          nullable: true
        receivedDate:
          type: string
          format: date
          nullable: true
        amount:
          type: number
          nullable: true
        status:
          type: string
          enum: [expected, received, overdue]
    RoyaltyCalcResult:
      type: object
      required: [contractId, calculations]
      properties:
        contractId:
          type: string
        partnerName:
          type: string
        royaltyType:
          type: string
          nullable: true
        royaltyDetails:
          type: string
          nullable: true
        calculations:
          type: array
          items:
            type: object
            required: [reportId, period, revenueAmount, amountOwed, reviewStatus]
            properties:
              reportId:
                type: string
              period:
                type: string
              revenueAmount:
                type: number
                nullable: true
              sharePercentage:
                type: number
                nullable: true
              amountOwed:
                type: number
                nullable: true
              reviewStatus:
                type: string
                enum: [pending, reviewed, approved]
              reviewedBy:
                type: string
                nullable: true
              reviewedAt:
                type: string
                format: date-time
                nullable: true
    UpdateRoyaltyStatusRequest:
      type: object
      required: [reportId, status]
      properties:
        reportId:
          type: string
        status:
          type: string
          enum: [pending, reviewed, approved]
...
    PaginatedRevenueReports:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RevenueReport'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: List all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, pageSize]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
    post:
      operationId: createUser
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}:
    get:
      operationId: getUser
      summary: Get user by ID (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: updateUser
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      operationId: deleteUser
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /contracts/{id}/revenue-reports:
    get:
      operationId: listContractRevenueReports
      summary: List revenue reports for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Revenue reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RevenueReport'
    post:
      operationId: createRevenueReport
      summary: Add a revenue report to a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRevenueReportRequest'
      responses:
        '201':
          description: Created revenue report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueReport'
  /revenue-reports:
    get:
      operationId: listRevenueReports
      summary: List all revenue reports (Finance/Admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [expected, received, overdue]
        - name: contractId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated revenue reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRevenueReports'
  /revenue-reports/{id}:
    put:
      operationId: updateRevenueReport
      summary: Update a revenue report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRevenueReportRequest'
      responses:
        '200':
          description: Updated revenue report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueReport'
    delete:
      operationId: deleteRevenueReport
      summary: Delete a revenue report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /royalties/{contractId}:
    get:
      operationId: getRoyaltyCalc
      summary: Get royalty calculations for a contract (Finance/Admin)
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Royalty calculation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoyaltyCalcResult'
  /royalties/{contractId}/approve:
    post:
      operationId: approveRoyalty
      summary: Approve or review a royalty calculation (Finance/Admin)
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoyaltyStatusRequest'
      responses:
        '200':
          description: Updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoyaltyCalcResult'
...